+++ src/routes/index.tsx
@@ -98,6 +98,7 @@ export default function Router() {
         },
         {
           path: "semester",
+          element: <Semester />,
           children: [{ path: "create", element: <CreateSemester /> }],
         },
       ],
@@ -130,6 +131,7 @@ const StudentType = Loadable(
 const CreateStudentType = Loadable(
   lazy(() => import("../pages/Settings/CreateStudentType"))
 );
+const Semester = Loadable(lazy(() => import("../pages/Semester/Semester")));
 const CreateSemester = Loadable(
   lazy(() => import("../pages/Semester/CreateSemester"))
 );